+++ src/main/java/io/github/eappezo/soundary/services/music/infrastructure/persistence/dao/JpaSharedMusicRepository.java
@@ -1,4 +1,4 @@
-package io.github.eappezo.soundary.services.music.infrastructure.persistence.repository;
+package io.github.eappezo.soundary.services.music.infrastructure.persistence.dao;
 
 import io.github.eappezo.soundary.services.music.infrastructure.persistence.SharedMusicEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
