+++ src/main/java/io/github/eappezo/soundary/services/music/infrastructure/persistence/dao/JpaTrackRepository.java
@@ -1,4 +1,4 @@
-package io.github.eappezo.soundary.services.music.infrastructure.persistence.repository;
+package io.github.eappezo.soundary.services.music.infrastructure.persistence.dao;
 
 import io.github.eappezo.soundary.services.music.infrastructure.persistence.TrackEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
