+++ src/main/java/io/github/eappezo/soundary/services/music/infrastructure/persistence/dao/SharedMusicRepositoryImpl.java
@@ -1,4 +1,4 @@
-package io.github.eappezo.soundary.services.music.infrastructure.persistence.repository;
+package io.github.eappezo.soundary.services.music.infrastructure.persistence.dao;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
